Covid-19 Data Exploration Project using SQL
The COVID-19 pandemic has had a profound impact on individuals, communities, and societies around the globe. The effects of the pandemic have been both short-term and long-term, with implications for human health, society, economy, and the environment. The outbreak has revealed weaknesses in global health systems, underlining the need for better preparation and coordination in managing pandemics. While the development and rollout of vaccines have provided hope for controlling the spread of the virus and lessening its impact, continued vigilance and cooperation are crucial to prevent future outbreaks and ensure the health and safety of individuals and communities. At present, the far-reaching consequences of COVID-19 are extensive and have significantly impacted daily life.

This project aims to delve into COVID-19 data using SQL with MSSQL, sourced from Our World in Data for the period between January 1, 2020, and April 10, 2023.

SUMMARY OF DATA FINDINGS 

As of April 10, 2023, the COVID-19 pandemic has resulted in 762,588,872 confirmed cases worldwide, with a recorded 6,895,830 deaths, indicating a global death rate of 0.90 percent. The highest number of confirmed cases were reported in the USA, China, India, France, and Germany, while the highest number of deaths were reported in the USA, Brazil, India, Russia, and Mexico. Among the continents, Asia had the highest number of confirmed cases, followed by Europe, North America, South America, Oceania, and Africa. The total number of deaths recorded by continent was highest in Europe, followed by North America, Asia, and South America. Vaccinations commenced on December 3, 2020. To visually represent the insights and trends derived from this analysis, I constructed a Power BI dashboard report.